requirements.txt
Flask==2.0.1
Flask-JWT-Extended==4.2.3
Flask-CORS==3.0.10
python-dotenv==0.19.0
libsql-client==0.3.0
SQLAlchemy==1.4.46